% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/are_reviews.R
\name{are_reviews}
\alias{are_reviews}
\title{Are the files reviews?}
\usage{
are_reviews(filenames)
}
\arguments{
\item{filenames}{names of the files}
}
\value{
a vector of TRUE and FALSE
}
\description{
Are the files reviews?
}
\author{
Richel Bilderbeek
}

