% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_formation.R
\name{check_formation}
\alias{check_formation}
\title{Check if the formation is valid}
\usage{
check_formation(formation, na_is_ok = FALSE)
}
\arguments{
\item{formation}{the K3 formation, in three, alphabettically
ordered uppercase characters, e.g \code{HKM} for the third
formation}

\item{na_is_ok}{boolean to indicate if \link{NA} is accepted}
}
\description{
Check if the formation is valid
}
\examples{
library(testthat)

expect_silent(check_formation("KKK"))
expect_silent(check_formation("JKK"))
expect_silent(check_formation("HKM"))
expect_silent(check_formation(NA, na_is_ok = TRUE))
expect_error(check_formation(NA, na_is_ok = FALSE))
expect_error(check_formation("3JS"))
expect_error(check_formation(""))
expect_error(check_formation(NULL))
expect_error(check_formation(3.14))
}
