% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_files.R
\name{collect_files}
\alias{collect_files}
\title{Collects all the filenames in inst/extdata}
\usage{
collect_files()
}
\value{
all filenames in inst/exdata
}
\description{
Collects all the filenames in inst/extdata
}
\examples{
  filenames <- collect_files()
  testit::assert(length(filenames) > 200)
}
\author{
Richel Bilderbeek
}

