% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_reviews.R
\name{collect_reviews}
\alias{collect_reviews}
\title{Collects all the reviews' filenames}
\usage{
collect_reviews()
}
\value{
all reviews' filenames
}
\description{
Collects all the reviews' filenames
}
\examples{
  filenames <- collect_reviews()
  testit::assert(length(filenames) > 100)
  testit::assert(is_review(filenames[1]))
}
\author{
Richel Bilderbeek
}

