% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_params_doc.R
\name{default_params_doc}
\alias{default_params_doc}
\title{This function does nothing. It is intended to inherit is parameters'
documentation.}
\usage{
default_params_doc(
  reviewer_name,
  song_name,
  song_filenames,
  song_filename,
  verbose
)
}
\arguments{
\item{reviewer_name}{name of the reviewer, e.g. \code{Richel}}

\item{song_name}{name of the K3 song}

\item{song_filenames}{paths to one or more song's review files}

\item{song_filename}{path to a song's review file}

\item{verbose}{set to TRUE for more output}
}
\description{
This function does nothing. It is intended to inherit is parameters'
documentation.
}
\note{
This is an internal function, so it should be marked with
\code{@noRd}. This is not done, as this will disallow all
functions to find the documentation parameters
}
\author{
Richèl J.C. Bilderbeek
}
