% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_paths.R
\name{find_path}
\alias{find_path}
\title{Finds the full path of a file}
\usage{
find_path(filename)
}
\arguments{
\item{filename}{the name of a file}
}
\value{
the full path of the filename if an existing file could be found, stops otherwise
}
\description{
Finds the full path of a file
}
\examples{
  path <- find_path("HeyahMama.md")
  testit::assert(file.exists(path))
}
\author{
Richel Bilderbeek
}

