% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_discography.R
\name{get_discography}
\alias{get_discography}
\title{Obtain the discography of K3}
\usage{
get_discography()
}
\value{
a datafrm
}
\description{
Obtain the discography of K3
}
\author{
Richel Bilderbeek
}

