% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_formation.R
\name{get_formation}
\alias{get_formation}
\title{Get the formation for a song}
\usage{
get_formation(song_filename)
}
\arguments{
\item{song_filename}{path to a song's review file}
}
\description{
Get the formation for a song
}
