% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_marks_rating.R
\name{get_marks_rating}
\alias{get_marks_rating}
\title{Get Mark's rating}
\usage{
get_marks_rating(song_filename)
}
\arguments{
\item{song_filename}{path to a song's review file}
}
\description{
Get Mark's rating
}
