% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rating.R
\name{get_rating}
\alias{get_rating}
\title{Extract the rating from a song's filename}
\usage{
get_rating(reviewer_name, song_filename)
}
\arguments{
\item{reviewer_name}{name of the reviewer, e.g. \code{Richel}}

\item{song_filename}{path to a song's review file}
}
\description{
Extract the rating from a song's filename
}
