% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ratings.R
\name{get_ratings}
\alias{get_ratings}
\title{Get ratings of multiple song filenames}
\usage{
get_ratings(song_filenames)
}
\arguments{
\item{song_filenames}{paths to one or more song's review files}
}
\description{
Get ratings of multiple song filenames
}
\examples{
library(testthat)

df <- get_ratings(get_song_path("10000Luchtballonnen.md"))
expect_true("filename" \%in\% names(df))
expect_true("formation" \%in\% names(df))
expect_true("rating" \%in\% names(df))
expect_true("reviewer" \%in\% names(df))
expect_equal(2, nrow(df))
}
