% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_richels_rating.R
\name{get_richels_rating}
\alias{get_richels_rating}
\title{Get Richel's rating}
\usage{
get_richels_rating(song_filename)
}
\arguments{
\item{song_filename}{path to a song's review file}
}
\description{
Get Richel's rating
}
