% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_song_path.R
\name{get_song_path}
\alias{get_song_path}
\title{Get the path of a song's name}
\usage{
get_song_path(song_name)
}
\arguments{
\item{song_name}{name of the K3 song}
}
\description{
Get the path of a song's name
}
