% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_song_rating.R
\name{get_song_rating}
\alias{get_song_rating}
\title{Get the rating from a song's filename
for a specific reviewer and formation}
\usage{
get_song_rating(song_filename, reviewer_name, formation)
}
\arguments{
\item{song_filename}{path to a song's review file}

\item{reviewer_name}{name of the reviewer, e.g. \code{Richel}}
}
\value{
a number if there is a rating given
to that formation by that reviewer,
else \link{NA}.
}
\description{
Get the rating from a song's filename
for a specific reviewer and formation
}
\examples{
library(testthat)

song_filename = get_song_path("Verliefd.md")
expect_equal(
  9,
  get_song_rating(
    reviewer_name = "Richel",
    song_filename = song_filename,
    formation = "KKK"
  )
)
expect_true(
  is.na(
    get_song_rating(
      reviewer_name = "Richel",
      song_filename = song_filename,
      formation = "JKK"
    )
  )
)
expect_equal(
  8,
  get_song_rating(
    reviewer_name = "Richel",
    song_filename = song_filename,
    formation = "HKM"
  )
)
expect_true(
  is.na(
    get_song_rating(
      reviewer_name = "Mark",
      song_filename = song_filename,
      formation = "KKK"
    )
  )
)
}
