% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_song_ratings.R
\name{get_song_ratings}
\alias{get_song_ratings}
\title{Get all the ratings from one song's filename}
\usage{
get_song_ratings(song_filename)
}
\arguments{
\item{song_filename}{path to a song's review file}
}
\description{
Get all the ratings from one song's filename
}
\examples{
library(testthat)

df <- get_song_ratings(
  song_filename = get_song_path("10000Luchtballonnen.md")
)
expect_true("formation" \%in\% names(df))
expect_true("rating" \%in\% names(df))
expect_true("reviewer" \%in\% names(df))
}
\seealso{
use \link{get_song_rating} to get the ratings
specifically for a reviewer and formation
}
