% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_songs_ratings.R
\name{get_songs_ratings}
\alias{get_songs_ratings}
\title{Get ratings of multiple song filenames}
\usage{
get_songs_ratings(song_filenames)
}
\arguments{
\item{song_filenames}{paths to one or more song's review files}
}
\description{
Get ratings of multiple song filenames
}
\examples{

df <- get_songs_ratings(
  song_filenames = c(
    get_song_path("10000Luchtballonnen.md"),
    get_song_path("10000Luchtballonnen.md")
  )
)
"filename" \%in\% names(df))
"formation" \%in\% names(df))
"rating" \%in\% names(df))
"reviewer" \%in\% names(df))
}
