% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_review.R
\name{is_review}
\alias{is_review}
\title{Is the file a review?}
\usage{
is_review(filename)
}
\arguments{
\item{filename}{name of the file}
}
\value{
a dataframe
}
\description{
Is the file a review?
}
\author{
Richel Bilderbeek
}

