% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_formation.R
\name{check_formation}
\alias{check_formation}
\title{Check if the formation is valid}
\usage{
check_formation(formation, na_is_ok = FALSE)
}
\arguments{
\item{formation}{the K3 formation, in three, alphabettically
ordered uppercase characters, e.g \code{HKM} for the third
formation}

\item{na_is_ok}{boolean to indicate if \link{NA} is accepted}
}
\description{
Check if the formation is valid
}
\examples{
check_formation("KKK")
check_formation("JKK")
check_formation("HKM")
check_formation(NA, na_is_ok = TRUE)
}
